old-tests/submission/W3C/domhtml/HTMLFormElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - FORM</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fs-action" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form accept-charset="US-ASCII" action="./files/getData.pl" enctype="application/x-www-form-urlencoded" id="form1" method="post">
<p>
<textarea cols="20" name="text1" rows="7"></textarea>
<input name="submit1" type="submit" value="Submit">
<input name="submit2" type="reset" value="Reset">
</p>
</form>



<script>
var domtest = async_test("The action attribute specifies the server-side form handler.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vaction;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("form");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vaction = testNode.action;

      assert_uri_equals(document.location.protocol,null,document.location.host,"getData.pl",null,null,null,true,vaction);

    domtest.done();
}));

</script>
<p>The action attribute specifies the server-side form handler. 

    Retrieve the action attribute and examine its value.</p>
<div id="log"></div>
</body></html>
